Add option to not inject distributed tracing headers
In this commit I'm modifying the middleware to take options in parameter.

First the options will need to be a Keyword list from now on (this could be revisited)
Second the options so far can take inject_distributed_tracing_headers option to propagate or not otel headers. By default we always propagate the otel headers.